Double -> double
git-svn-id: https://svn.science.uu.nl/repos/project.fisme.java/AlgebraPijlenOpdr@5170 58dd2db4-0f11-0410-9df4-da15f00f50bd

+++ src/fi/algebrapijlenopdr/AftrekSchuifComponent.java
@@ -61,8 +61,8 @@ public class AftrekSchuifComponent extends BewerkingSchuifComponent
 			{	d = e1.kind2.geefWaarde().doubleValue() - e2.geefWaarde().doubleValue();
 //System.out.println("e1 + d = " + d);			
 			}
-			else if (e1 instanceof Aftrekking && (Double.isNaN(e1.kind1.geefWaarde()) || 
-					 e1.kind1.geefWaarde() != 0))
+			else if (e1 instanceof Aftrekking && (Double.isNaN(e1.kind1.geefWaarde().doubleValue()) || 
+					 e1.kind1.geefWaarde().doubleValue() != 0))
 			{	d = -e1.kind2.geefWaarde().doubleValue() - e2.geefWaarde().doubleValue();
 //System.out.println("e1 - d = " + d);			
 			}
